+++ bbs/msgbase.cpp
@@ -21,7 +21,9 @@
 #include "bbs/inmsg.h"
 #include "bbs/input.h"
 #include "bbs/keycodes.h"
+#include "bbs/message_file.h"
 #include "bbs/netsup.h"
+#include "bbs/read_message.h"
 #include "bbs/wconstants.h"
 #include "bbs/wstatus.h"
 #include "bbs/bbs.h"
@@ -42,282 +44,6 @@ using std::string;
 using std::stringstream;
 using std::unique_ptr;
 using namespace wwiv::strings;
-
-// Local function prototypes
-File* OpenMessageFile(const string messageAreaFileName);
-void set_gat_section(File *pMessageFile, int section);
-void save_gat(File *pMessageFile);
-
-static long gat_section;
-
-/**
- * Sets the global variables pszOutOriginStr and pszOutOriginStr2.
- * Note: This is a private function
- */
-static void SetMessageOriginInfo(int nSystemNumber, int nUserNumber, string* outNetworkName,
-                                 string* outLocation) {
-  string netName;
-
-  if (session()->GetMaxNetworkNumber() > 1) {
-    netName = StrCat(net_networks[session()->GetNetworkNumber()].name, "- ");
-  }
-
-  outNetworkName->clear();
-  outLocation->clear();
-
-  if (IsEqualsIgnoreCase(session()->GetNetworkName(), "Internet") ||
-      nSystemNumber == 32767) {
-    outNetworkName->assign("Internet Mail and Newsgroups");
-    return;
-  }
-
-  if (nSystemNumber && session()->GetCurrentNetworkType() == net_type_wwivnet) {
-    net_system_list_rec *csne = next_system(nSystemNumber);
-    if (csne) {
-      string netstatus;
-      if (nUserNumber == 1) {
-        if (csne->other & other_net_coord) {
-          netstatus = "{NC}";
-        } else if (csne->other & other_group_coord) {
-          netstatus = StringPrintf("{GC%d}", csne->group);
-        } else if (csne->other & other_area_coord) {
-          netstatus = "{AC}";
-        }
-      }
-      const string filename = StringPrintf(
-              "%s%s%c%s.%-3u",
-              syscfg.datadir,
-              REGIONS_DIR,
-              File::pathSeparatorChar,
-              REGIONS_DIR,
-              atoi(csne->phone));
-
-      string description;
-      if (File::Exists(filename)) {
-        // Try to use the town first.
-        const string phone_prefix = StringPrintf("%c%c%c", csne->phone[4], csne->phone[5], csne->phone[6]);
-        description = describe_area_code_prefix(atoi(csne->phone), atoi(phone_prefix.c_str()));
-      }
-      if (description.empty()) {
-        // Try area code if we still don't have a description.
-        description = describe_area_code(atoi(csne->phone));
-      }
-
-      *outNetworkName = StrCat(netName, csne->name, " [", csne->phone, "] ", netstatus.c_str());
-      *outLocation = (!description.empty()) ? description : "Unknown Area";
-    } else {
-      *outNetworkName = StrCat(netName, "Unknown System");
-      *outLocation = "Unknown Area";
-    }
-  }
-}
-
-/**
- * Deletes a message
- * This is a public function.
- */
-void remove_link(messagerec * pMessageRecord, string fileName) {
-  switch (pMessageRecord->storage_type) {
-  case 0:
-  case 1:
-    break;
-  case 2: {
-    unique_ptr<File> file(OpenMessageFile(fileName));
-    if (file->IsOpen()) {
-      set_gat_section(file.get(), static_cast<int>(pMessageRecord->stored_as / GAT_NUMBER_ELEMENTS));
-      long lCurrentSection = pMessageRecord->stored_as % GAT_NUMBER_ELEMENTS;
-      while (lCurrentSection > 0 && lCurrentSection < GAT_NUMBER_ELEMENTS) {
-        long lNextSection = static_cast<long>(gat[lCurrentSection]);
-        gat[lCurrentSection] = 0;
-        lCurrentSection = lNextSection;
-      }
-      save_gat(file.get());
-      file->Close();
-    }
-  }
-  break;
-  default:
-    // illegal storage type
-    break;
-  }
-}
-
-/**
- * Opens the message area file {pszMessageAreaFileName} and returns the file handle.
- * Note: This is a Private method to this module.
- */
-File* OpenMessageFile(const string messageAreaFileName) {
-  application()->GetStatusManager()->RefreshStatusCache();
-
-  const string filename = StrCat(syscfg.msgsdir, messageAreaFileName, FILENAME_DAT_EXTENSION);
-  File *pFileMessage = new File(filename);
-  if (!pFileMessage->Open(File::modeReadWrite | File::modeBinary)) {
-    // Create message area file if it doesn't exist.
-    pFileMessage->Open(File::modeBinary | File::modeCreateFile | File::modeReadWrite);
-    for (int i = 0; i < GAT_NUMBER_ELEMENTS; i++) {
-      gat[i] = 0;
-    }
-    pFileMessage->Write(gat, GAT_SECTION_SIZE);
-    pFileMessage->SetLength(GAT_SECTION_SIZE + (75L * 1024L));
-    gat_section = 0;
-  }
-  pFileMessage->Seek(0L, File::seekBegin);
-  pFileMessage->Read(gat, GAT_SECTION_SIZE);
-
-  gat_section = 0;
-  return pFileMessage;
-}
-
-#define GATSECLEN (GAT_SECTION_SIZE + GAT_NUMBER_ELEMENTS * MSG_BLOCK_SIZE)
-#ifndef MSG_STARTING
-#define MSG_STARTING (gat_section * GATSECLEN + GAT_SECTION_SIZE)
-#endif  // MSG_STARTING
-
-long current_gat_section() { return gat_section; }
-void current_gat_section(long section) { gat_section = section; }
-
-void set_gat_section(File *pMessageFile, int section) {
-  if (gat_section != section) {
-    long lFileSize = pMessageFile->GetLength();
-    long lSectionPos = static_cast<long>(section) * GATSECLEN;
-    if (lFileSize < lSectionPos) {
-      pMessageFile->SetLength(lSectionPos);
-      lFileSize = lSectionPos;
-    }
-    pMessageFile->Seek(lSectionPos, File::seekBegin);
-    if (lFileSize < (lSectionPos + GAT_SECTION_SIZE)) {
-      for (int i = 0; i < GAT_NUMBER_ELEMENTS; i++) {
-        gat[i] = 0;
-      }
-      pMessageFile->Write(gat, GAT_SECTION_SIZE);
-    } else {
-      pMessageFile->Read(gat, GAT_SECTION_SIZE);
-    }
-    gat_section = section;
-  }
-}
-
-void save_gat(File *pMessageFile) {
-  long lSectionPos = static_cast<long>(gat_section) * GATSECLEN;
-  pMessageFile->Seek(lSectionPos, File::seekBegin);
-  pMessageFile->Write(gat, GAT_SECTION_SIZE);
-  WStatus *pStatus = application()->GetStatusManager()->BeginTransaction();
-  pStatus->IncrementFileChangedFlag(WStatus::fileChangePosts);
-  application()->GetStatusManager()->CommitTransaction(pStatus);
-}
-
-void savefile(char *b, long lMessageLength, messagerec * pMessageRecord, const string fileName) {
-  WWIV_ASSERT(pMessageRecord);
-  switch (pMessageRecord->storage_type) {
-  case 0:
-  case 1:
-    break;
-  case 2: {
-    int gati[128];
-    unique_ptr<File> pMessageFile(OpenMessageFile(fileName));
-    if (pMessageFile->IsOpen()) {
-      for (int section = 0; section < 1024; section++) {
-        set_gat_section(pMessageFile.get(), section);
-        int gatp = 0;
-        int nNumBlocksRequired = static_cast<int>((lMessageLength + 511L) / MSG_BLOCK_SIZE);
-        int i4 = 1;
-        while (gatp < nNumBlocksRequired && i4 < GAT_NUMBER_ELEMENTS) {
-          if (gat[i4] == 0) {
-            gati[gatp++] = i4;
-          }
-          ++i4;
-        }
-        if (gatp >= nNumBlocksRequired) {
-          gati[gatp] = -1;
-          for (int i = 0; i < nNumBlocksRequired; i++) {
-            pMessageFile->Seek(MSG_STARTING + MSG_BLOCK_SIZE * static_cast<long>(gati[i]), File::seekBegin);
-            pMessageFile->Write((&b[i * MSG_BLOCK_SIZE]), MSG_BLOCK_SIZE);
-            gat[gati[i]] = static_cast< unsigned short >(gati[i + 1]);
-          }
-          save_gat(pMessageFile.get());
-          break;
-        }
-      }
-      pMessageFile->Close();
-    }
-    pMessageRecord->stored_as = static_cast<long>(gati[0]) + static_cast<long>(gat_section) * GAT_NUMBER_ELEMENTS;
-  }
-  break;
-  default: {
-    bout.bprintf("WWIV:ERROR:msgbase.cpp: Save - storage_type=%u!\r\n", pMessageRecord->storage_type);
-    WWIV_ASSERT(false);
-  }
-  break;
-  }
-}
-
-char *readfile(messagerec * pMessageRecord, string fileName, long *plMessageLength) {
-  *plMessageLength = 0L;
-  if (pMessageRecord->storage_type == 2) {
-    unique_ptr<File> file(OpenMessageFile(fileName));
-    set_gat_section(file.get(), pMessageRecord->stored_as / GAT_NUMBER_ELEMENTS);
-    int lCurrentSection = pMessageRecord->stored_as % GAT_NUMBER_ELEMENTS;
-    long lMessageLength = 0;
-    while (lCurrentSection > 0 && lCurrentSection < GAT_NUMBER_ELEMENTS) {
-      lMessageLength += MSG_BLOCK_SIZE;
-      lCurrentSection = gat[lCurrentSection];
-    }
-    if (lMessageLength == 0) {
-      bout << "\r\nNo message found.\r\n\n";
-      return nullptr;
-    }
-    char* b = new char[lMessageLength + 512];
-    if (!b) {
-      return nullptr;
-    }
-    lCurrentSection = pMessageRecord->stored_as % GAT_NUMBER_ELEMENTS;
-    long lMessageBytesRead = 0;
-    while (lCurrentSection > 0 && lCurrentSection < GAT_NUMBER_ELEMENTS) {
-      file->Seek(MSG_STARTING + MSG_BLOCK_SIZE * static_cast< long >(lCurrentSection), File::seekBegin);
-      lMessageBytesRead += static_cast<long>(file->Read(&(b[lMessageBytesRead]), MSG_BLOCK_SIZE));
-      lCurrentSection = gat[lCurrentSection];
-    }
-    file->Close();
-    long lRealMessageLength = lMessageBytesRead - MSG_BLOCK_SIZE;
-    while ((lRealMessageLength < lMessageBytesRead) && (b[lRealMessageLength] != CZ)) {
-      ++lRealMessageLength;
-    }
-    *plMessageLength = lRealMessageLength;
-    b[lRealMessageLength + 1] = '\0';
-    return b;
-  }
-  return nullptr;
-}
-
-void LoadFileIntoWorkspace(const char *pszFileName, bool bNoEditAllowed) {
-  File fileOrig(pszFileName);
-  if (!fileOrig.Open(File::modeBinary | File::modeReadOnly)) {
-    bout << "\r\nFile not found.\r\n\n";
-    return;
-  }
-
-  long lOrigSize = fileOrig.GetLength();
-  unique_ptr<char[]> b(new char[lOrigSize + 1024]);
-  fileOrig.Read(b.get(), lOrigSize);
-  fileOrig.Close();
-  if (b[lOrigSize - 1] != CZ) {
-    b[lOrigSize++] = CZ;
-  }
-
-  File fileOut(syscfgovr.tempdir, INPUT_MSG);
-  fileOut.Open(File::modeBinary | File::modeCreateFile | File::modeReadWrite);
-  fileOut.Write(b.get(), lOrigSize);
-  fileOut.Close();
-
-  use_workspace = (bNoEditAllowed || !okfsed()) ? true : false;
-
-  if (!session()->IsNewMailWatiting()) {
-    bout << "\r\nFile loaded into workspace.\r\n\n";
-    if (!use_workspace) {
-      bout << "Editing will be allowed.\r\n";
-    }
-  }
-}
 
 // returns true on success (i.e. the message gets forwarded)
 bool ForwardMessage(int *pUserNumber, int *pSystemNumber) {
@@ -786,8 +512,7 @@ void email(int nUserNumber, int nSystemNumber, bool forceit, int anony, bool for
   messageRecord.storage_type = EMAIL_STORAGE;
   int nUseFSED = (bAllowFSED) ? INMSG_FSED : INMSG_NOFSED;
   string title;
-  inmsg(&messageRecord, &title, &i, !forceit, "email", nUseFSED, szDestination, MSGED_FLAG_NONE, force_title);
-  if (messageRecord.stored_as == 0xffffffff) {
+  if (!inmsg(&messageRecord, &title, &i, !forceit, "email", nUseFSED, szDestination, MSGED_FLAG_NONE, force_title)) {
     return;
   }
   if (anony & anony_sender) {
@@ -961,381 +686,32 @@ void imail(int nUserNumber, int nSystemNumber) {
   }
 }
 
-void read_message1(messagerec * pMessageRecord, char an, bool readit, bool *next, const char *pszFileName,
-                   int nFromSystem, int nFromUser) {
-  string strName, strDate;
-  char s[205];
-  long lMessageTextLength;
-
-  string origin_str;
-  string origin_str2;
-
-  // Moved internally from outside this method
-  SetMessageOriginInfo(nFromSystem, nFromUser, &origin_str, &origin_str2);
-
-  g_flags &= ~g_flag_ansi_movement;
-
-  unique_ptr<char[]> ss;
-  bool ansi = false;
-  *next = false;
-  long lCurrentCharPointer = 0;
-  bool abort = false;
-  switch (pMessageRecord->storage_type) {
-  case 0:
-  case 1:
-  case 2: {
-    ss.reset(readfile(pMessageRecord, pszFileName, &lMessageTextLength));
-    if (ss == nullptr) {
-      bout.Color(6);
-      osan("File not found.", &abort, next);
-      bout.nl(2);
-      return;
-    }
-    int nNamePtr = 0;
-    while ((ss[nNamePtr] != RETURN) &&
-           (nNamePtr < lMessageTextLength) &&
-           (nNamePtr < 200)) {
-      strName.push_back(ss[nNamePtr++]);
-    }
-    ++nNamePtr;
-    int nDatePtr = 0;
-    if (ss[nNamePtr] == SOFTRETURN) {
-      ++nNamePtr;
-    }
-    while ((ss[nNamePtr + nDatePtr] != RETURN) &&
-           static_cast<long>(nNamePtr + nDatePtr < lMessageTextLength) &&
-           (nDatePtr < 60)) {
-      strDate.push_back(ss[(nDatePtr++) + nNamePtr]);
-    }
-
-    lCurrentCharPointer = nNamePtr + nDatePtr + 1;
-  }
-  break;
-  default:
-    // illegal storage type
-    bout.nl();
-    bout << "->ILLEGAL STORAGE TYPE<-\r\n\n";
+void LoadFileIntoWorkspace(const char *pszFileName, bool bNoEditAllowed) {
+  File fileOrig(pszFileName);
+  if (!fileOrig.Open(File::modeBinary | File::modeReadOnly)) {
+    bout << "\r\nFile not found.\r\n\n";
     return;
   }
 
-  irt_name[0] = '\0';
-  g_flags |= g_flag_disable_mci;
-  string name;
-  string date = strDate;
-  string from;
-  string loc;
-  switch (an) {
-  default:
-  case 0:
-    if (syscfg.sysconfig & sysconfig_enable_mci) {
-      g_flags &= ~g_flag_disable_mci;
-    }
-    strcpy(irt_name, strName.c_str());
-    name = strName;
-    from = origin_str;
-    loc = origin_str2;
-    break;
-  case anony_sender:
-    if (readit) {
-      name = StrCat("<<< ", strName, " >>>");
-    } else {
-      name = ">UNKNOWN<";
-    }
-    break;
-  case anony_sender_da:
-  case anony_sender_pp:
-    date = ">UNKNOWN<";
-    if (an == anony_sender_da) {
-      name = "Abby";
-    } else {
-      name = "Problemed Person";
-    }
-    if (readit) {
-      name = StrCat("<<< ", strName, " >>>");
-    }
-    break;
+  long lOrigSize = fileOrig.GetLength();
+  unique_ptr<char[]> b(new char[lOrigSize + 1024]);
+  fileOrig.Read(b.get(), lOrigSize);
+  fileOrig.Close();
+  if (b[lOrigSize - 1] != CZ) {
+    b[lOrigSize++] = CZ;
   }
 
-  bout << "|#9Name|#7: |#1" << name << wwiv::endl;
-  bout << "|#9Date|#7: |#1" << date << wwiv::endl;
-  if (!from.empty()) {
-    bout << "|#9From|#7: |#1" << from << wwiv::endl;
-  }
-  if (!loc.empty()) {
-    bout << "|#9Loc|#7:  |#1" << loc << wwiv::endl;
-  }
+  File fileOut(syscfgovr.tempdir, INPUT_MSG);
+  fileOut.Open(File::modeBinary | File::modeCreateFile | File::modeReadWrite);
+  fileOut.Write(b.get(), lOrigSize);
+  fileOut.Close();
 
-  int nNumCharsPtr  = 0;
-  int nLineLenPtr   = 0;
-  int ctrld     = 0;
-  char ch       = 0;
-  bool done     = false;
-  bool printit    = false;
-  bool ctrla      = false;
-  bool centre     = false;
+  use_workspace = (bNoEditAllowed || !okfsed()) ? true : false;
 
-  checka(&abort, next);
-  bout.nl();
-  while (!done && !abort && !hangup) {
-    switch (pMessageRecord->storage_type) {
-    case 0:
-    case 1:
-    case 2:
-      ch = ss[lCurrentCharPointer];
-      if (lCurrentCharPointer >= lMessageTextLength) {
-        ch = CZ;
-      }
-      break;
+  if (!session()->IsNewMailWatiting()) {
+    bout << "\r\nFile loaded into workspace.\r\n\n";
+    if (!use_workspace) {
+      bout << "Editing will be allowed.\r\n";
     }
-    if (ch == CZ) {
-      done = true;
-    } else {
-      if (ch != SOFTRETURN) {
-        if (ch == RETURN || !ch) {
-          printit = true;
-        } else if (ch == CA) {
-          ctrla = true;
-        } else if (ch == CB) {
-          centre = true;
-        } else if (ch == CD) {
-          ctrld = 1;
-        } else if (ctrld == 1) {
-          if (ch >= '0' && ch <= '9') {
-	    if ( ch == '0' ) {
-	      ctrld = -1; // don't display
-	    } else {
-	      if (session()->user()->GetOptionalVal() == 0 ) {
-		ctrld = 0; // display
-	      } else {
-		if (10 - (session()->user()->GetOptionalVal()) < ( ch - '0' )) {
-		  ctrld = -1; // don't display
-		} else {
-		  ctrld = 0; // display
-		}
-	      }
-	    }
-	  } else {
-            ctrld = 0; // ctrl-d and non-numeric
-          }
-        } else {
-          if (ch == ESC) {
-            ansi = true;
-          }
-          if (g_flags & g_flag_ansi_movement) {
-            g_flags &= ~g_flag_ansi_movement;
-            lines_listed = 0;
-            if (session()->localIO()->GetTopLine() && session()->localIO()->GetScreenBottom() == 24) {
-              session()->localIO()->set_protect(0);
-            }
-          }
-          s[nNumCharsPtr++] = ch;
-          if (ch == CC || ch == BACKSPACE) {
-            --nLineLenPtr;
-          } else {
-            ++nLineLenPtr;
-          }
-        }
-
-        if (printit || ansi || nLineLenPtr >= 80) {
-          if (centre && (ctrld != -1)) {
-            int nSpacesToCenter = (session()->user()->GetScreenChars() - session()->localIO()->WhereX() -
-                                   nLineLenPtr) / 2;
-            osan(charstr(nSpacesToCenter, ' '), &abort, next);
-          }
-          if (nNumCharsPtr) {
-            if (ctrld != -1) {
-              if ((session()->localIO()->WhereX() + nLineLenPtr >= session()->user()->GetScreenChars()) && !centre
-                  && !ansi) {
-                bout.nl();
-              }
-              s[nNumCharsPtr] = '\0';
-              osan(s, &abort, next);
-              if (ctrla && s[nNumCharsPtr - 1] != SPACE && !ansi) {
-                if (session()->localIO()->WhereX() < session()->user()->GetScreenChars() - 1) {
-                  bputch(SPACE);
-		  bout.nl();
-                } else {
-                  bout.nl();
-                }
-                checka(&abort, next);
-              }
-            }
-            nLineLenPtr  = 0;
-            nNumCharsPtr = 0;
-          }
-          centre = false;
-        }
-        if (ch == RETURN) {
-          if (ctrla == false) {
-            if (ctrld != -1) {
-              bout.nl();
-              checka(&abort, next);
-            }
-          } else {
-            ctrla = false;
-          }
-          if (printit) {
-            ctrld = 0;
-          }
-        }
-        printit = false;
-      } else {
-        ctrld = 0;
-      }
-    }
-    ++lCurrentCharPointer;
-  }
-  if (!abort && nNumCharsPtr) {
-    s[nNumCharsPtr] = '\0';
-    bout << s;
-    bout.nl();
-  }
-  bout.Color(0);
-  bout.nl();
-  if (express && abort && !*next) {
-    expressabort = true;
-  }
-  if (ansi && session()->topdata && session()->IsUserOnline()) {
-    application()->UpdateTopScreen();
-  }
-  if (syscfg.sysconfig & sysconfig_enable_mci) {
-    g_flags &= ~g_flag_disable_mci;
-  }
-}
-
-void read_message(int n, bool *next, int *val) {
-  bout.nl();
-  bool abort = false;
-  *next = false;
-  if (session()->user()->IsUseClearScreen()) {
-    bout.cls();
-  }
-  if (forcescansub) {
-    bout.cls();
-    bout.GotoXY(1, 1);
-    bout << "|#4   FORCED SCAN OF SYSOP INFORMATION - YOU MAY NOT ABORT.  PLEASE READ THESE!  |#0\r\n";
-  }
-
-  bout.bprintf(" |#9Msg|#7: [|#1%u|#7/|#1%lu|#7]|#%d |#2%s\r\n", n,
-                                    session()->GetNumMessagesInCurrentMessageArea(), session()->GetMessageColor(),
-                                    subboards[session()->GetCurrentReadMessageArea()].name);
-  const string subjectLine = "|#9Subj|#7: ";
-  osan(subjectLine, &abort, next);
-  bout.Color(session()->GetMessageColor());
-  postrec p = *get_post(n);
-  if (p.status & (status_unvalidated | status_delete)) {
-    bout.Color(6);
-    osan("<<< NOT VALIDATED YET >>>", &abort, next);
-    bout.nl();
-    if (!lcs()) {
-      return;
-    }
-    *val |= 1;
-    osan(subjectLine, &abort, next);
-    bout.Color(session()->GetMessageColor());
-  }
-  strncpy(irt, p.title, 60);
-  irt_name[0] = '\0';
-  bout.Color(session()->GetMessageColor());
-  osan(irt, &abort, next);
-  bout.nl();
-  if ((p.status & status_no_delete) && lcs()) {
-    osan("|#9Info|#7: ", &abort, next);
-    bout.Color(session()->GetMessageColor());
-    osan("Permanent Message", &abort, next);
-    bout.nl();
-  }
-  if ((p.status & status_pending_net) && session()->user()->GetSl() > syscfg.newusersl) {
-    osan("|#9Val|#7:  ", &abort, next);
-    bout.Color(session()->GetMessageColor());
-    osan("Not Network Validated", &abort, next);
-    bout.nl();
-    *val |= 2;
-  }
-  if (!abort) {
-    bool bReadit = (lcs() || (getslrec(session()->GetEffectiveSl()).ability & ability_read_post_anony)) ? true : false;
-    int nNetNumSaved = session()->GetNetworkNumber();
-
-    if (p.status & status_post_new_net) {
-      set_net_num(p.network.network_msg.net_number);
-    }
-    read_message1(&(p.msg), static_cast<char>(p.anony & 0x0f), bReadit, next,
-                  (subboards[session()->GetCurrentReadMessageArea()].filename), p.ownersys, p.owneruser);
-
-    if (nNetNumSaved != session()->GetNetworkNumber()) {
-      set_net_num(nNetNumSaved);
-    }
-
-    session()->user()->SetNumMessagesRead(session()->user()->GetNumMessagesRead() + 1);
-    session()->SetNumMessagesReadThisLogon(session()->GetNumMessagesReadThisLogon() + 1);
-  } else if (express && !*next) {
-    expressabort = true;
-  }
-  if (p.qscan > qsc_p[session()->GetCurrentReadMessageArea()]) {
-    qsc_p[session()->GetCurrentReadMessageArea()] = p.qscan;
-  }
-  
-  unsigned long current_qscan_pointer = 0;
-  {
-    std::unique_ptr<WStatus> wwiv_status(application()->GetStatusManager()->GetStatus());
-    // not sure why we check this twice...
-    // maybe we need a getCachedQScanPointer?
-    current_qscan_pointer = wwiv_status->GetQScanPointer();
-  }
-  if (p.qscan >= current_qscan_pointer) {
-    WStatus* pStatus = application()->GetStatusManager()->BeginTransaction();
-    if (p.qscan >= pStatus->GetQScanPointer()) {
-      pStatus->SetQScanPointer(p.qscan + 1);
-    }
-    application()->GetStatusManager()->CommitTransaction(pStatus);
-  }
-}
-
-void lineadd(messagerec* pMessageRecord, const string& sx, string fileName) {
-  const string line = StringPrintf("%s\r\n\x1a", sx.c_str());
-
-  switch (pMessageRecord->storage_type) {
-  case 0:
-  case 1:
-    break;
-  case 2: {
-    unique_ptr<File> message_file(OpenMessageFile(fileName));
-    set_gat_section(message_file.get(), pMessageRecord->stored_as / GAT_NUMBER_ELEMENTS);
-    int new1 = 1;
-    while (new1 < GAT_NUMBER_ELEMENTS && gat[new1] != 0) {
-      ++new1;
-    }
-    int i = static_cast<int>(pMessageRecord->stored_as % GAT_NUMBER_ELEMENTS);
-    while (gat[i] != 65535) {
-      i = gat[i];
-    }
-    char *b = nullptr;
-    if ((b = static_cast<char*>(BbsAllocA(GAT_NUMBER_ELEMENTS))) == nullptr) {
-      message_file->Close();
-      return;
-    }
-    message_file->Seek(MSG_STARTING + static_cast<long>(i) * MSG_BLOCK_SIZE, File::seekBegin);
-    message_file->Read(b, MSG_BLOCK_SIZE);
-    int j = 0;
-    while (j < MSG_BLOCK_SIZE && b[j] != CZ) {
-      ++j;
-    }
-    strcpy(&(b[j]), line.c_str());
-    message_file->Seek(MSG_STARTING + static_cast<long>(i) * MSG_BLOCK_SIZE, File::seekBegin);
-    message_file->Write(b, MSG_BLOCK_SIZE);
-    if (((j + line.size()) > MSG_BLOCK_SIZE) && (new1 != GAT_NUMBER_ELEMENTS)) {
-      message_file->Seek(MSG_STARTING + static_cast<long>(new1)  * MSG_BLOCK_SIZE, File::seekBegin);
-      message_file->Write(b + MSG_BLOCK_SIZE, MSG_BLOCK_SIZE);
-      gat[new1] = 65535;
-      gat[i] = static_cast< unsigned short >(new1);
-      save_gat(message_file.get());
-    }
-    free(b);
-    message_file->Close();
-  }
-  break;
-  default:
-    // illegal storage type
-    break;
   }
 }
