+++ bbs/quote.cpp
@@ -25,6 +25,7 @@
 #include "bbs/printfile.h"
 #include "bbs/bbs.h"
 #include "bbs/fcns.h"
+#include "bbs/message_file.h"
 #include "bbs/vars.h"
 #include "core/strings.h"
 #include "core/textfile.h"
